chore(http-cache): split api-platform/http-cache (#5531)
---------
