+++ composer.json
@@ -120,7 +120,10 @@
     "autoload": {
         "psr-4": {
             "ApiPlatform\\": "src/"
-        }
+        },
+        "files": [
+            "src/deprecation.php"
+        ]
     },
     "autoload-dev": {
         "psr-4": {
