+++ src/HttpCache/.gitignore
@@ -1,0 +1,3 @@
+/composer.lock
+/vendor
+/.phpunit.result.cache
