+++ src/HttpCache/README.md
@@ -1,0 +1,7 @@
+# API Platform - HTTP Cache
+
+HTTP Cache support.
+
+## Resources
+
+
