+++ src/HttpCache/composer.json
@@ -1,0 +1,66 @@
+{
+  "name": "api-platform/http-cache",
+  "description": "HttpCache support",
+  "type": "library",
+  "keywords": [
+    "Cache",
+    "Http"
+  ],
+  "homepage": "https://api-platform.com",
+  "license": "MIT",
+  "authors": [
+    {
+      "name": "Kévin Dunglas",
+      "email": "kevin@dunglas.fr",
+      "homepage": "https://dunglas.fr"
+    },
+    {
+      "name": "API Platform Community",
+      "homepage": "https://api-platform.com/comunnity/contributors"
+    }
+  ],
+  "require": {
+    "php": ">=8.1",
+    "api-platform/metadata": "*@dev || ^3.1",
+    "symfony/http-foundation": "^6.1"
+  },
+  "require-dev": {
+    "guzzlehttp/guzzle": "^6.0 || ^7.0",
+    "symfony/dependency-injection": "^6.1",
+    "phpspec/prophecy-phpunit": "^2.0",
+    "symfony/phpunit-bridge": "^6.1",
+    "symfony/http-client": "^6.1"
+  },
+  "autoload": {
+    "psr-4": {
+      "ApiPlatform\\HttpCache\\": ""
+    },
+    "exclude-from-classmap": [
+      "/Tests/"
+    ]
+  },
+  "config": {
+    "preferred-install": {
+      "*": "dist"
+    },
+    "sort-packages": true,
+    "allow-plugins": {
+      "composer/package-versions-deprecated": true,
+      "phpstan/extension-installer": true
+    }
+  },
+  "extra": {
+    "branch-alias": {
+      "dev-main": "3.2.x-dev"
+    },
+    "symfony": {
+      "require": "^6.1"
+    }
+  },
+  "repositories": [
+    {
+      "type": "path",
+      "url": "../Metadata"
+    }
+  ]
+}
