+++ src/Symfony/EventListener/AddHeadersListener.php
@@ -11,7 +11,7 @@
 
 declare(strict_types=1);
 
-namespace ApiPlatform\HttpCache\EventListener;
+namespace ApiPlatform\Symfony\EventListener;
 
 use ApiPlatform\Metadata\Resource\Factory\ResourceMetadataCollectionFactoryInterface;
 use ApiPlatform\Util\OperationRequestInitiatorTrait;
@@ -22,8 +22,6 @@ use Symfony\Component\HttpKernel\Event\ResponseEvent;
  * Configures cache HTTP headers for the current response.
  *
  * @author Kévin Dunglas <dunglas@gmail.com>
- *
- * @deprecated use \Symfony\EventListener\AddHeadersListener.php instead
  */
 final class AddHeadersListener
 {
